update: Notification 도메인을 Support 레이어로 리팩터링
