--- src/main/java/team/washer/server/v2/domain/notification/service/SendFcmNotificationService.java
@@ -1,8 +1,0 @@
-package team.washer.server.v2.domain.notification.service;
-
-import team.washer.server.v2.domain.user.entity.User;
-
-public interface SendFcmNotificationService {
-
-    void execute(User user, String title, String body);
-}
